% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/classify_default.R
\docType{data}
\name{classify_default}
\alias{classify_default}
\title{classify_default}
\description{
classify_default
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

