% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getChangeDist.R
\name{getChangeDist}
\alias{getChangeDist}
\title{getChangeDist}
\usage{
getChangeDist(sample, point=rep(0,nrow(sample)), base=sample[1,], margin=1, tol=0.1)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{point}{An optional numeric vector containing the location of the disruption (e.g. the mutation in an RNA)}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wildtype SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{tol}{An optional number indicating the tolerance for the change. Default is 0.1.}
}
\value{
A numeric vector of change distances.
}
\description{
A function to get the average distance of change from the disruption site
}
\details{
This function determines the average distance of change from the disruption site

This function calculates the average distance of change from the disruption using the change in pattern to determine the location of changes.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get change distance
loc = getChangeDist(samp_nreduce)
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getFeatures}}
}
\keyword{RNA}
\keyword{change}
\keyword{distance}

