% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getChangeParams.R
\name{getChangeParams}
\alias{getChangeParams}
\title{getChangeParams}
\usage{
getChangeParams(sample, base=NULL, margin=1, \cr trim=0, high=NULL, tol=0.1, outfile=NULL, append=F)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wild type SHAPE trace). Default is the first trace in each file.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{trim}{An optional number indicating the number of nucleotides to be trimed from the ends. Default is 0.}

\item{high}{An optional number indicating the reactivity above which reactivities are considered high. Default is third quartile of the sample in each file.}

\item{tol}{An optional number indicating the tolerance for the change. Default is 0.1.}

\item{outfile}{An optional string indicating the name of the output file. The output file will consist of two columns (magnitude change and pattern change). Default will not output a file.}

\item{append}{An optional boolean to append the file if an outfile is given. Default is FALSE.}

\item{outfile}{An optional output file for the table.}
}
\value{
A two column numeric matrix for magnitude and pattern change.
}
\description{
A function to get magnitude and pattern change parameters
}
\details{
This calculates the magnitude and pattern change in SHAPE reactivity.

This function normalizes and reduces the noise in the sample. The magnitude and pattern change is calculated for the sample using the magnitudeChange and patternChange functions.
}
\examples{
#input files
data("shape_ex")
#get change parameters
params = getChangeParams(shape_ex, trim=5, outfile="out.txt")
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{magnitudeChange}} \code{\link{patternChange}} \code{\link{normalize}} \code{\link{reduceNoise}} \code{\link{classifyRNA}} \code{\link{predict.classifyRNA}} \code{\link{getExampleData}}
}
\keyword{RNA}
\keyword{change}
\keyword{magnitude}
\keyword{parameters}
\keyword{pattern}

