% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChangeRange.R
\name{getChangeRange}
\alias{getChangeRange}
\title{getChangeRange}
\usage{
getChangeRange(sample, base=sample[1,], margin=1, tol=0.1)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wildtype SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{tol}{An optional number indicating the tolerance for the change. Default is 0.1.}
}
\value{
A numeric vector of trace change ranges.
}
\description{
A function to get the range of change for a SHAPE trace.
}
\details{
This function calculates the range of change positions for each row (or column) in sample.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get change range
cr = getChangeRange(samp_nreduce)

}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getFeatures}}
}
\keyword{RNA}
\keyword{change}
\keyword{range}
\keyword{trace}

