% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContiguous.R
\name{getContiguous}
\alias{getContiguous}
\title{getContiguous}
\usage{
getContiguous(sample, base=sample[1,], margin=1, tol=0.1)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wildtype SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{tol}{An optional number indicating the tolerance for the change. Default is 0.1.}
}
\value{
A numeric vector of counts
}
\description{
A function to get the number of stretches of contiguous change.
}
\details{
This function calculates the number of stretches of contiguous change between the base vector and each row (or column) in sample.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get trace difference
contig = getContiguous(samp_norm)
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getFeatures}}
}
\keyword{RNA}
\keyword{getContiguous}
\keyword{trace}

