% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getExampleData.R
\name{getExampleData}
\alias{getExampleData}
\title{getExampleData}
\usage{
getExampleData(name)
}
\arguments{
\item{name}{A string indicating which data set to return}
}
\value{
A numeric matrix for mutmap, magpat_ex, or shape_ex.
}
\description{
A function to load data for use in the classifyRNA package
}
\details{
This function returning the data sets available for the classifyRNA package

This function returns data from mutmap.Rdata, magpat_ex.Rdata or shape_ex.Rdata
}
\note{
The data has been gathered from the RNA Mapping Database mutate and map experiments.
}
\examples{
obj = getExampleData("mutmap")
}
\author{
Chanin Tolson
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\seealso{
\code{\link{classifyRNA}} \code{\link{getChangeParams}} \code{\link{mutmap}} \code{\link{magpat_ex}} \code{\link{shape_ex}}
}
\keyword{RNA}
\keyword{data}
\keyword{example}

