% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{getFeatures}
\usage{
getFeatures(sample, base=NULL, margin=1,  norm=T, noise=T, trim=0, high=NULL,
   tol=0.1, point=rep(0,nrow(sample)), window=ncol(sample), outfile=NULL, append=F)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wild type SHAPE trace). Default is the first trace in each file.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{norm}{An optional boolean to normalize the sample. Default is TRUE.}

\item{noise}{An optional boolean to reduce noise in the sample. Default is TRUE.}

\item{trim}{An optional number indicating the number of nucleotides to be trimed from the ends. Default is 0.}

\item{high}{An optional number indicating the reactivity above which reactivities are considered high. Default is third quartile of the sample in each file.}

\item{tol}{An optional number indicating the tolerance for the change. Default is 0.1.}

\item{point}{An optional numerical vector indicating the location of disruption (e.g. mutation point)}

\item{window}{An optional number indicating the number of columns around the disruption to calculate. Default is the entire trace.}

\item{outfile}{An optional string indicating the name of the output file. The output file will consist of two columns (magnitude change and pattern change). Default will not output a file.}

\item{append}{An optional boolean to append the file if an outfile is given. Default is FALSE.}
}
\value{
\describe{
 \item{"outmat"}{A three column numeric matrix for magnitude, pattern, location and timewarp change.}
 \item{"outfile"}{An optional output file for the matrix.}
}
}
\description{
A function to get the features for describing RNA structure change. These features can be used in classification of RNA structure change.
}
\details{
This calculates the change in SHAPE reactivity traces.

This function calculates the magnitude correlation coefficient, pattern  correlation coefficient, average change distance, dynamic time warping, average trace difference and rna length.
}
\examples{
#input files
data("shape_ex")
#get features
params = getFeatures(shape_ex, trim=5, outfile="out.txt")
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{normalize}} \code{\link{reduceNoise}} \code{\link{getMagCC}} \code{\link{getPatternCC}} \code{\link{getChangeDist}} \code{\link{getTimeWarping}} \code{\link{getTraceDiff}}
}
\keyword{RNA}
\keyword{change}
\keyword{features}
\keyword{structure}

