% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getL2norm.R
\name{getL2norm}
\alias{getL2norm}
\title{getL2norm}
\usage{
getL2norm(sample, base=sample[1,], margin=1)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wildtype SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}
}
\value{
A numeric vector of L2 norm values.
}
\description{
A function to get the L2 norm.
}
\details{
This function calculates the L2 norm between the base vector and each row (or column) in sample.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get trace difference
l2norm = getL2norm(samp_nreduce)
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getFeatures}}
}
\keyword{L2}
\keyword{RNA}
\keyword{norm}
\keyword{trace}

