% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPatternCC.R
\name{getPatternCC}
\alias{getPatternCC}
\title{getPatternCC}
\usage{
getPatternCC(sample, base=sample[1,], margin=1, tol=0.1)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the value to which the samples are to be compared (e.g. a wild type SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{tol}{An optional number indicating the tolerance for the change. Default is 0.1.}
}
\value{
A numeric vector of pattern correlation coefficients.
}
\description{
A function to get the correlation coefficient between SHAPE trace change patterns.
}
\details{
The pattern for a single SHAPE reacitivity trace is the pattern of increase in reactivity or decrease in reactivity between nucleotides. If the change is less than the tolerance value, it is considered a none change. The pattern change value is the Pearson correlation coefficient between the base vector pattern and the pattern of each row (or column) in sample.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get pattern correlation coefficient
pat = getPatternCC(samp_nreduce)

}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getFeatures}}
}
\keyword{RNA}
\keyword{correlation-coefficient}
\keyword{pattern}

