% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/locationChange.R
\name{locationChange}
\alias{locationChange}
\title{locationChange}
\usage{
locationChange(sample, point=rep(0,nrow(sample)), base=sample[1,], margin=1)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{point}{An optional numeric vector containing the location of the disruption (e.g. the mutation in an RNA)}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wildtype SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}
}
\value{
A numeric vector of location changes.
}
\description{
A function to get the location change between samples
}
\details{
This function determines

This function calculates the distance of change from the disruption using changepoint analysis to determine the location of change.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get location change
loc = locationChange(samp_nreduce)
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getChangeParams}}
}
\keyword{RNA}
\keyword{change}
\keyword{location}

