% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutmap-data.R
\docType{data}
\name{mutmap}
\alias{mutmap}
\title{Consensus classification, magnitude change, and pattern change in example RNAs}
\format{An object of class \code{data.frame} with 2670 rows and 11 columns.}
\usage{
mutmap
}
\description{
Example dataset with the calculated magnitude correlation coefficient, pattern correlation coefficient, average change distance, dynamic time warping, and average trace difference for a set of RNAs. Features calculated from SHAPE traces from the RNA Mapping Database. Parameters calculated using getFeatures function in classSNitch package. Consensus classification determined using crowd-sourced manual classification.
}
\examples{
data("mutmap")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

