% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(sample, base=sample[1,], margin=1, outbase=FALSE, mean=1.5)
}
\arguments{
\item{sample}{A numeric matrix containing values to be normalized (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the sample is to be normalized (e.g. a wild type SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if sample is organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{outbase}{An optional boolean indicating if the normalized base should be returned. Default is FALSE.}

\item{mean}{An optional number setting the mean SHAPE value. Default is 1.5.}
}
\value{
\describe{
 \item{"samp_norm"}{A normalized numeric matrix with the same dimensions as sample.} 
 \item{"samp_norm"}{An optional list with two elements: normalized numeric matrix with the same dimensions as sample and a normalized vector the same length as base.}
}
}
\description{
A between-sample normalization function for SHAPE traces.
}
\details{
This function normalizes the average value of the base vector. Each row (or column) in sample is then normalized by minimizing the absolute difference between the base and the sample row (or column).
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
           
#sample data
sample = matrix(sample(1:100), ncol=10)
base = sample(1:100, size=10)
#normalize
samp_norm = normalize(sample, base)

}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getFeatures}}
}
\keyword{RNA}
\keyword{between-sample}
\keyword{normalize}

