% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/predict.classifyRNA.R
\name{predict.classifyRNA}
\alias{predict.classifyRNA}
\title{predict.classifyRNA}
\usage{
predict(cr, sample, type="response")
}
\arguments{
\item{cr}{A classifyRNA object (see classifyRNA function).}

\item{sample}{A matrix of predictors for magnitude and pattern change (e.g. output from getChangeParams())}

\item{type}{A string indicating what type of random forest data to return ("response", "vote", "prob"). Default is "response".}
}
\value{
A numeric matrix of "response", "vote" or "prob" predictions.
}
\description{
A function to classify rna structure change
}
\details{
This function predicts RNA structure change in SHAPE data

This function predicts RNA structure change in SHAPE data using a random forest classifier.
}
\note{
Options for type:
\itemize{
 \item{"response"}{ Predicted classes (classes with majority vote)}
 \item{"vote"}{ Vote count fraction (one column for each class and one row for each input)}
 \item{"class"}{ Class probabilities (one column for each class and one row for each input)}
}
}
\examples{
#input data
sample = getExampleData("magpat_ex")
#build classifier
cr = classifyRNA(classes=2)
#get prediction
cr_pred = predict(cr, sample, type="vote")
}
\author{
Chanin Tolson
}
\references{
A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. R News 2(3), 18--22 (randomForest package)
}
\seealso{
\code{\link{getChangeParams}} \code{\link{classifyRNA}} \code{\link{getExampleData}}
}
\keyword{RNA}
\keyword{change}
\keyword{predict}
\keyword{prediction}
\keyword{structure}

