% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_ex-data.R
\docType{data}
\name{shape_ex}
\alias{shape_ex}
\title{16S rRNA four-way junction wild-type and mutant SHAPE data}
\format{An object of class \code{matrix} with 111 rows and 110 columns.}
\usage{
shape_ex
}
\description{
Example SHAPE traces from 16SFWJ_1M7_0001 Mutate and Map experiment in the RNA Mapping Database.
}
\examples{
data("shape_ex")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

