% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/timewarpChange.R
\name{timewarpChange}
\alias{timewarpChange}
\title{timewarpChange}
\usage{
timewarpChange(sample, base=sample[1,], margin=1)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wildtype SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}
}
\value{
A numeric vector of timewarp changes.
}
\description{
A function to get the timewarp between samples
}
\details{
This function compares the timewarp between samples.

This function calculates the timewarp between the base vector and each row (or column) in sample.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get magnitude change
tw = timewarpChange(samp_nreduce)
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getChangeParams}}
}
\keyword{RNA}
\keyword{change}
\keyword{timewarp}

