% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/traceChange.R
\name{traceChange}
\alias{traceChange}
\title{traceChange}
\usage{
traceChange(sample, base=sample[1,], margin=1,
   point=rep(0,nrow(sample)), window=ncol(sample))
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wildtype SHAPE trace). Default is the first trace in sample.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{point}{An optional numeric vector containing the location of the disruption (e.g. the mutation in an RNA)}

\item{window}{An optional number indicating the number of columns around the disruption to calculate. Default is the entire trace.}
}
\value{
A numeric vector of trace changes.
}
\description{
A function to get the average trace between samples
}
\details{
This function compares the average trace difference between samples.

This function calculates the average trace difference between the base vector and each row (or column) in sample.
}
\examples{
#sample data
sample = matrix(sample(1:100), ncol=10)
#normalize
samp_norm = normalize(sample)
#reduce noise
samp_nreduce = reduceNoise(samp_norm, trim=1, high=4)
#get trace change
tc = traceChange(samp_nreduce)
}
\author{
Chanin Tolson
}
\seealso{
\code{\link{getChangeParams}}
}
\keyword{RNA}
\keyword{change}
\keyword{trace}

