% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_default.R
\docType{data}
\name{classify_default}
\alias{classify_default}
\title{classify_default}
\description{
A dataset of class labels and features SHAPE-traces from the mutate and map experiments found in the RMDB
}
\note{
Organization of the data file: header=TRUE, tab-delimited .rda file
\itemize{
 \item{"column 1"}{ class label} 
 \item{"column 2"}{ pattern correlation} 
 \item{"column 3"}{ dynamic time warping} 
 \item{"column 4"}{ contiguousness} 
 \item{"column 5"}{ magnitude correlation} 
 \item{"column 6"}{ change variance} 
 \item{"column 7"}{ eSDC} 
 \item{"column 8"}{ change range} 
 \item{"column 9"}{ L2 norm} 
}
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

