% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_default.R
\docType{data}
\name{classify_default}
\alias{classify_default}
\title{classify_default}
\description{
A dataset of class labels and change features from SHAPE-traces in the RMDB
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

