% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{getFeatures}
\usage{
getFeatures(sample, base=NULL, margin=1,  norm=T, noise=T, mean=1.5, trim=0, high=NULL,
   tol=0.1, outfile=NULL, append=F)
}
\arguments{
\item{sample}{A numeric matrix containing values to be compared (e.g. a set of mutant SHAPE traces).}

\item{base}{An optional numeric vector containing the values to which the samples are to be compared (e.g. a wild type SHAPE trace). Default is the first trace in each file.}

\item{margin}{An optional number indicating if the samples are organized by rows or columns, where 1 indicates rows and 2 indicates columns. Default is 1.}

\item{norm}{An optional boolean to normalize the sample. Default is TRUE.}

\item{noise}{An optional boolean to reduce noise in the sample. Default is TRUE.}

\item{mean}{An optional number setting the mean SHAPE value for normalization. Default is 1.5.}

\item{trim}{An optional number indicating the number of nucleotides to be trimed from the ends. Default is 0.}

\item{high}{An optional number indicating the reactivity above which reactivities are considered high. Default is third quartile of the sample in each file.}

\item{tol}{An optional number indicating the tolerance for the change. Default is 0.1.}

\item{outfile}{An optional string indicating the name of the output file. The output file will consist of feature columns. Default will not output a file.}

\item{append}{An optional boolean to append the file if an outfile is given. Default is FALSE.}
}
\value{
\describe{
 \item{"outmat"}{A seven column numeric matrix for pattern correlation coefficient, dynamic time warping, contiguousness of change, magnitude correlation coefficient, change of variance, experimental structural disruption coefficient, change range, and L2 norm} 
 \item{"outfile"}{An optional output file for the matrix.}
}
}
\description{
A function to get the features for describing RNA structure change. These features can be used in classification of RNA structure change.
}
\details{
This function calculates the  pattern correlation coefficient, dynamic time warping, contiguousness of change, magnitude correlation coefficient, change of variance, experimental structural disruption coefficient, change range and L2 norm. These are features used to describe structure change in SHAPE traces.
}
\examples{
#input files
data("shape_ex")
#get features
params = getFeatures(shape_ex, trim=5, outfile="out.txt")

}
\author{
Chanin Tolson
}
\seealso{
\code{\link{normalize}} \code{\link{reduceNoise}} \code{\link{getPatternCC}} \code{\link{getTimeWarping}} \code{\link{getContiguous}} \code{\link{getMagCC}} \code{\link{getESDC}} \code{\link{getChangeRange}} \code{\link{getL2norm}}
}
\keyword{RNA}
\keyword{change}
\keyword{features}
\keyword{structure}

