% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/magpat_ex-data.R
\docType{data}
\name{magpat_ex}
\alias{magpat_ex}
\title{Magnitude and Pattern Change in example RNAs}
\format{\preformatted{'data.frame':	2670 obs. of  5 variables:
 $ class           : int  1 1 NA 1 1 1 NA NA NA 1 ...
 $ magnitude.change: num  0.994 0.995 0.995 0.984 0.988 ...
 $ pattern.change  : num  0.861 0.929 0.912 0.922 0.904 ...
 $ location.change : num  54.8 91.2 68.8 68.2 80.3 ...
 $ t[, 4]          : num  25.9 21.2 22.4 26.1 24.1 ...
}}
\usage{
magpat_ex
}
\description{
Example dataset with the calculated magnitude, pattern, location and timewarp change for a set of RNAs. SHAPE traces from the RNA Mapping Database. Parameters calculated using getChangeParameters function.
}
\examples{
data("magpat_ex")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

