% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/magpat_ex-data.R
\docType{data}
\name{magpat_ex}
\alias{magpat_ex}
\title{Magnitude and Pattern Change in example RNAs}
\format{\preformatted{'data.frame':	2670 obs. of  6 variables:
 $ class           : int  1 1 NA 1 1 1 NA NA NA 1 ...
 $ magnitude.change: num  0.0947 0.0949 0.0949 0.0938 0.0942 ...
 $ pattern.change  : num  0.0821 0.0886 0.087 0.0879 0.0862 ...
 $ location.change : num  5.95 5.97 5.65 3.67 4.71 ...
 $ timewarp.change : num  2.47 2.02 2.13 2.49 2.3 ...
 $ trace.change    : num  2.52 2.81 3.42 3.6 4.23 ...
}}
\usage{
magpat_ex
}
\description{
Example dataset with the calculated magnitude, pattern, location and timewarp change for a set of RNAs. SHAPE traces from the RNA Mapping Database. Parameters calculated using getChangeParameters function.
}
\examples{
data("magpat_ex")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

