% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutmap-data.R
\docType{data}
\name{mutmap}
\alias{mutmap}
\title{Consensus classification and features for example RNA SHAPE traces. Features calculated from SHAPE traces from the RNA Mapping Database. Parameters calculated using getFeatures function in classSNitch package. Consensus classification determined using crowd-sourced manual classification.}
\format{An object of class \code{matrix} with 167 rows and 9 columns.}
\usage{
mutmap
}
\description{
Consensus classification and features for example RNA SHAPE traces. Features calculated from SHAPE traces from the RNA Mapping Database. Parameters calculated using getFeatures function in classSNitch package. Consensus classification determined using crowd-sourced manual classification.
}
\examples{
data("mutmap")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

