% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mutmap-data.R
\docType{data}
\name{mutmap}
\alias{mutmap}
\title{Consensus classification, magnitude change, and pattern change in example RNAs}
\format{\preformatted{'data.frame':	2670 obs. of  9 variables:
 $ NvLG.response   : num  0 0 NA 0 0 0 NA NA NA 0 ...
 $ NLvG.response   : int  0 0 NA 0 0 0 NA NA NA 0 ...
 $ LvG.response    : int  NA NA NA NA NA NA NA NA NA NA ...
 $ magnitude change: num  0.988 0.993 0.984 0.978 0.988 ...
 $ pattern change  : num  0.82 0.888 0.866 0.894 0.899 ...
 $ change distance : num  0.559 0.633 0.557 0.441 0.482 ...
 $ time warping    : num  0.142 0.114 0.141 0.154 0.113 ...
 $ trace difference: num  2.52 2.45 1.95 2.97 4.23 ...
 $ length          : num  110 110 110 110 110 110 110 110 110 110 ...
}}
\usage{
mutmap
}
\description{
Example dataset with the calculated magnitude correlation coefficient, pattern correlation coefficient, average change distance, dynamic time warping, and average trace difference for a set of RNAs. Features calculated from SHAPE traces from the RNA Mapping Database. Parameters calculated using getFeatures function in classSNitch package. Consensus classification determined using crowd-sourced manual classification.
}
\examples{
data("mutmap")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

