% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/mutmap-data.R
\docType{data}
\name{mutmap}
\alias{mutmap}
\title{Consensus classification, magnitude change, and pattern change in example RNAs}
\format{\preformatted{'data.frame':	2670 obs. of  11 variables:
 $ NvLvG.Response  : int  1 1 NA 1 1 1 NA NA NA 1 ...
 $ NvLG.Response   : int  1 1 NA 1 1 1 NA NA NA 1 ...
 $ GvNL.Response   : int  0 0 NA 0 0 0 NA NA NA 0 ...
 $ LvG.Response    : int  NA NA NA NA NA NA NA NA NA NA ...
 $ NvLG.Votes      : num  1 NA 0.75 1 1 NA NA NA 0.75 NA ...
 $ GvNL.Votes      : num  0 0 NA 0 0 0 NA NA NA 0 ...
 $ LvG.Votes       : num  NA NA NA NA NA NA NA NA NA NA ...
 $ magnitude.change: num  0.0947 0.0949 0.0949 0.0938 0.0942 ...
 $ pattern.change  : num  0.0821 0.0886 0.087 0.0879 0.0862 ...
 $ location.change : num  0 0 0 0 0 0 0 0 0 0 ...
 $ timewarp.change : num  2.47 2.02 2.13 2.49 2.3 ...
}}
\usage{
mutmap
}
\description{
Example dataset with the calculated magnitude, pattern, location and timewarp change for a set of RNAs. SHAPE traces from the RNA Mapping Database. Parameters calculated using getChangeParameters function. Consensus classification determined using crowd-sourced manual classification.
}
\examples{
data("mutmap")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

