% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.classifyRNA.R
\name{predict.classifyRNA}
\alias{predict}
\alias{predict.classifyRNA}
\title{predict.classifyRNA}
\usage{
\method{predict}{classifyRNA}(object, sample = NULL, resp="prob", ...)
}
\arguments{
\item{object}{An object of classifyRNA (see classifyRNA function).}

\item{sample}{An optional matrix of predictors (e.g. output from getFeatures())}

\item{resp}{An optional string to determine type of return value. Default is "prob".}

\item{...}{additional arguments for predict method}
}
\value{
A matrix of "response", "vote" or "prob" predictions.
\describe{
 \item{"response"}{ Predicted classes (classes with majority vote)}
 \item{"vote"}{ Vote count fraction (one column for each class and one row for each input)} 
 \item{"prob"}{ Class probabilities (one column for each class and one row for each input)}
}
}
\description{
A function to classify rna structure change from an existing classifier.
}
\details{
This function predicts RNA structure change in SHAPE data using a random forest classifier.
}
\examples{
#input data
data("mutmap")
#build classifier
cr = classifyRNA()
#get prediction
cr_pred = predict(cr, mutmap[,2:9])

}
\author{
Chanin Tolson
}
\references{
A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. R News 2(3), 18--22 (randomForest package)
}
\seealso{
\code{\link{getFeatures}} \code{\link{classifyRNA}}
}
\keyword{RNA}
\keyword{change}
\keyword{predict}
\keyword{prediction}
\keyword{structure}

