% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/predict.classifyRNA.R
\name{predict.classifyRNA}
\alias{predict.classifyRNA}
\title{predict.classifyRNA}
\usage{
predict(object, sample=NULL, ...)
}
\arguments{
\item{object}{An object of classifyRNA (see classifyRNA function).}

\item{sample}{An optional matrix of predictors for magnitude, pattern, location and pattern change (e.g. output from getChangeParams())}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of "response", "vote" or "prob" predictions.
\describe{
 \item{"response"}{ Predicted classes (classes with majority vote)}
 \item{"vote"}{ Vote count fraction (one column for each class and one row for each input)}
 \item{"class"}{ Class probabilities (one column for each class and one row for each input)}
}
}
\description{
A function to classify rna structure change
}
\details{
This function predicts RNA structure change in SHAPE data

This function predicts RNA structure change in SHAPE data using a random forest classifier.
}
\examples{
#input data
data("magpat_ex")
#build classifier
cr = classifyRNA(classes=2)
#get prediction
cr_pred = predict(cr, magpat_ex[,2:5])
}
\author{
Chanin Tolson
}
\references{
A. Liaw and M. Wiener (2002). Classification and Regression by randomForest. R News 2(3), 18--22 (randomForest package)
}
\seealso{
\code{\link{getChangeParams}} \code{\link{classifyRNA}}
}
\keyword{RNA}
\keyword{change}
\keyword{predict}
\keyword{prediction}
\keyword{structure}

