% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_ex-data.R
\docType{data}
\name{shape_ex}
\alias{shape_ex}
\title{16S rRNA four-way junction wild-type and mutant SHAPE data from mutate and map experiments in the RMDB.}
\format{An object of class \code{matrix} with 111 rows and 110 columns.}
\usage{
shape_ex
}
\description{
16S rRNA four-way junction wild-type and mutant SHAPE data from mutate and map experiments in the RMDB.
}
\examples{
data("shape_ex")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

