% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/shape_ex-data.R
\docType{data}
\name{shape_ex}
\alias{shape_ex}
\title{16S rRNA four-way junction wild-type and mutant SHAPE data}
\format{\preformatted{ num [1:111, 1:110] 18.89 11.57 3.69 13.7 10.68 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:111] "16SFWJ_1M7_0001_WT" "16SFWJ_1M7_0001_G126C" "16SFWJ_1M7_0001_G127C" "16SFWJ_1M7_0001_G128C" ...
  ..$ : chr [1:110] "V3" "V4" "V5" "V6" ...
}}
\usage{
shape_ex
}
\description{
Example SHAPE traces from 16SFWJ_1M7_0001 Mutate and Map experiment in the RNA Mapping Database.
}
\examples{
data("shape_ex")
}
\references{
\href{http://rmdb.stanford.edu/}{RNA Mapping Database}
}
\keyword{datasets}

